Remove switch from low level interpreter
The big switch based interpreter is not really efficient, as a lot of
boundary checks are performed for fetching and decoding the
instructions.
Having an approach more related to the way computed gotos work allows
for a better execution time. The only boundary check for an interpretation
step is now performed for fetching the instruction.